all-in-one-downloader/requirements.txt
Flask
yt-dlp
gallery-dl
razorpay
firebase-admin
google-cloud-firestore
python-dotenv
